requirements.txt
django
mysqlclient
djoser
djangorestframework-simplejwt
pillow